Fix CMempoolTxDB addition and removal to check for duplicate or missing txns.
The method CMempoolTxDB::AddTransactions does not check whether any of the
transactions being added are already in the DB. This does matter to the
underlying levelDB, but if duplicate txns are added using this method then
the disk usage and txn count accounting values calculated by this class will
be wrong. This commit changes the behaviour of this method to only update
the accounting information for transactions that are not clready in the DB.

Also make corresponding changes to CMempoolTxDB::RemoveTransactions to fix
the same issue there.

Note that the documentation for this class specifies it should only be used
from a single threaded context.